requirements.txt
python_telegram_bot
pymongo~=3.11.3
django-environ
